# internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: Fleetsweeper API
  version: 0.1.0
  description: |
    HTTP API for the Fleetsweeper server. Read endpoints are unauthenticated
    when --insecure is set; mutating endpoints require a bearer token unless
    --insecure is in effect.
  license:
    name: MIT
    url: https://github.com/dcadolph/fleetsweeper/blob/main/LICENSE
  contact:
    name: Fleetsweeper project
    url: https://github.com/dcadolph/fleetsweeper
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://{host}
    description: Operator-deployed server
    variables:
      host:
        default: fleet.internal
security:
  - bearerAuth: []
  - {}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token configured via --auth-token. Required for POST/DELETE.
  schemas:
...
    ScanRecord:
      type: object
      required: [id, timestamp, clusters]
      properties:
        id: {type: string, example: "01HF8Z..."}
        timestamp: {type: string, format: date-time}
        clusters:
          type: array
          items: {type: string}
        scanners:
          type: array
          items: {type: string}
...
paths:
...
  /api/scans:
    get:
      summary: List recent scans
      parameters:
        - in: query
          name: limit
          schema: {type: integer, default: 50}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/ScanRecord"}
    post:
      summary: Trigger a new scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                contexts:
                  type: array
                  items: {type: string}
                all_contexts: {type: boolean}
                group: {type: string}
      responses:
        "202":
          description: Scan accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  scan_id: {type: string}
                  status: {type: string}
        "429":
          description: A scan is already in progress
...